create_fixtures.py: merge create_fixtures_aux_calcs.py and create_fixtures.py
