+++ tests/fixtures/create_fixtures.py
@@ -19,25 +19,144 @@ Options:
 '''
 
 import docopt
+import numpy as np
+import mpmath
 import sys
 
 import lif_meanfield_tools as lmt
+from lif_meanfield_tools.meanfield_calcs import (
+    effective_coupling_strength,
+    )
+from lif_meanfield_tools.aux_calcs import (
+    d_nu_d_mu,
+    d_nu_d_mu_fb433,
+    d_nu_d_nu_in_fb,
+    Phi,
+    Phi_prime_mu,
+    Psi,
+    d_Psi,
+    d_2_Psi,
+    p_hat_boxcar
+    )
+
 ureg = lmt.ureg
 
 
-def fixture_working_point(network):
-    network.working_point()
+def fix_additional_rates_for_fixed_input(network, file):
+    nu_e_ext, nu_i_ext = network.additional_rates_for_fixed_input(
+        network.network_params['mean_input_set'],
+        network.network_params['std_input_set'])
+    network.results['add_nu_e_ext'] = nu_e_ext
+    network.results['add_nu_i_ext'] = nu_i_ext
+    network.save(file_name=file)
+    
+
+def fix_d_nu_d_mu(network, file):
+    params = network.network_params
+    mus = network.mean_input()
+    sigmas = network.std_input()
+    results = [d_nu_d_mu(params['tau_m'],
+                         params['tau_r'],
+                         params['V_th_rel'],
+                         params['V_0_rel'],
+                         mu, sigma)
+               for mu, sigma in zip(mus, sigmas)]
+    network.results['d_nu_d_mu'] = results
+    network.save(file_name=file)
 
 
-def fixture_transfer_function(network):
-    """Calculate results for all options of transfer_function."""
-    network.transfer_function(method='shift')
-    network.results['tf_shift'] = network.results.pop('transfer_function')
-    network.transfer_function(method='taylor')
-    network.results['tf_taylor'] = network.results['transfer_function']
+def fix_d_nu_d_mu_fb433(network, file):
+    params = network.network_params
+    mus = network.mean_input()
+    sigmas = network.std_input()
+    results = [d_nu_d_mu_fb433(params['tau_m'],
+                               params['tau_s'],
+                               params['tau_r'],
+                               params['V_th_rel'],
+                               params['V_0_rel'],
+                               mu, sigma)
+               for mu, sigma in zip(mus, sigmas)]
+    network.results['d_nu_d_mu_fb433'] = results
+    network.save(file_name=file)
 
 
-def fixture_delay_dist_matrix(network):
+def fix_d_nu_d_nu_in_fb(network, file):
+    params = network.network_params
+    mus = network.mean_input()
+    sigmas = network.std_input()
+    results = [d_nu_d_nu_in_fb(params['tau_m'],
+                               params['tau_s'],
+                               params['tau_r'],
+                               params['V_th_rel'],
+                               params['V_0_rel'],
+                               params['j'],
+                               mu, sigma)
+               for mu, sigma in zip(mus, sigmas)]
+    network.results['d_nu_d_nu_in_fb'] = results
+    network.save(file_name=file)
+
+    
+def fix_d_Psi(fixture_path):
+    """Call d_Psi for a range of possible inputs and save result as fixture."""
+
+    function_name = 'd_Psi'
+    output_file = fixture_path + function_name + '.npz'
+    
+    z_range = np.concatenate([-np.logspace(2, -5, 4), [0],
+                              np.logspace(-5, 2, 4)])
+    a, b = np.meshgrid(z_range, z_range)
+    zs = a.flatten() + complex(0, 1) * b.flatten()
+    xs = np.linspace(-10, 10, 8)
+    
+    zs, xs = np.meshgrid(zs, xs)
+    zs = zs.flatten()
+    xs = xs.flatten()
+    
+    psi_outputs = []
+    for z, x in zip(zs, xs):
+        psi_outputs.append(Psi(z + 1, x))
+    
+    outputs = []
+    for z, x in zip(zs, xs):
+        outputs.append(d_Psi(z, x))
+    
+    np.savez(output_file, zs=zs, xs=xs, psis=psi_outputs, outputs=outputs)
+    
+    
+def fix_d_2_Psi(fixture_path):
+    """Call d_2_Psi for a range of inputs and save result as fixture."""
+
+    function_name = 'd_2_Psi'
+    output_file = fixture_path + function_name + '.npz'
+    
+    z_range = np.concatenate([-np.logspace(2, -5, 4), [0],
+                              np.logspace(-5, 2, 4)])
+    a, b = np.meshgrid(z_range, z_range)
+    zs = a.flatten() + complex(0, 1) * b.flatten()
+    xs = np.linspace(-10, 10, 8)
+    
+    zs, xs = np.meshgrid(zs, xs)
+    zs = zs.flatten()
+    xs = xs.flatten()
+    
+    psi_outputs = []
+    for z, x in zip(zs, xs):
+        psi_outputs.append(Psi(z + 2, x))
+        
+    outputs = []
+    for z, x in zip(zs, xs):
+        outputs.append(d_2_Psi(z, x))
+    
+    np.savez(output_file, zs=zs, xs=xs, psis=psi_outputs, outputs=outputs)
+    
+    
+def fix_delay_dist_single(network, file):
+    """Calculate fixtures for all delay dist matrix options."""
+    network.delay_dist_matrix(network.analysis_params['omega'])
+    network.save(file_name=file)
+    
+    
+def fix_delay_dist_matrix(network, file):
     """Calculate fixtures for all delay dist matrix options."""
     original_delay_dist = network.network_params['delay_dist']
     network.network_params['delay_dist'] = 'none'
@@ -53,9 +172,24 @@ def fixture_delay_dist_matrix(network):
     network.results['delay_dist_truncated_gaussian'] = dd_truncated_gaussian
     network.results['delay_dist_gaussian'] = dd_gaussian
     network.network_params['delay_dist'] = original_delay_dist
+    network.save(file_name=file)
+        
+
+def fix_eff_coupling_strength(network, file):
+    eff_coupling_strength = effective_coupling_strength(
+        network.network_params['tau_m'],
+        network.network_params['tau_s'],
+        network.network_params['tau_r'],
+        network.network_params['V_0_rel'],
+        network.network_params['V_th_rel'],
+        network.network_params['J'],
+        network.results['mean_input'],
+        network.results['std_input'])
+    network.results['effective_coupling_strength'] = eff_coupling_strength
+    network.save(file_name=file)
     
     
-def fixture_eigenspectra(network):
+def fix_eigenspectra(network, file):
     regime = network.network_params['regime']
     
     network.eigenvalue_spectra('MH')
@@ -76,78 +210,132 @@ def fixture_eigenspectra(network):
     if regime != 'negative_firing_rate':
         network.l_eigenvec_spectra('prop_inv')
     
+    network.save(file_name=file)
     
-def fixture_power_spectra(network):
+    
+def fix_p_hat_boxcar(fixture_path):
+    """Call p_hat_boxcar for a range of inputs and save result as fixture."""
+    
+    function_name = 'p_hat_boxcar'
+    output_file = fixture_path + function_name + '.npz'
+    
+    lp = -5
+    hp = 5
+    steps = 20
+    ks = np.concatenate([-np.logspace(hp, lp, steps), [0],
+                         np.logspace(lp, hp, steps)])
+
+    widths = np.logspace(-5, 5)
+    
+    ks, widths = np.meshgrid(ks, widths)
+    ks = ks.flatten()
+    widths = widths.flatten()
+    
+    outputs = []
+    
+    for k, width in zip(ks, widths):
+        outputs.append(p_hat_boxcar(k, width))
+        
+    np.savez(output_file, ks=ks, widths=widths, outputs=outputs)
+
+    
+def fix_Phi(fixture_path):
+    """Call Phi for a range of possible inputs and save result as fixture."""
+    
+    function_name = 'Phi'
+    output_file = fixture_path + function_name + '.npz'
+    
+    lp = -5
+    hp = 1.5
+    test_inputs = np.concatenate([-np.logspace(hp, lp),
+                                  [0],
+                                  np.logspace(lp, hp)])
+    
+    outputs = []
+    for test_input in test_inputs:
+        outputs.append(Phi(test_input))
+    
+    np.savez(output_file, s_values=test_inputs, outputs=outputs)
+
+
+def fix_Phi_prime_mu(fixture_path):
+    """Call Phi_prime_mu for a range of inputs and save result as fixture."""
+    
+    function_name = 'Phi_prime_mu'
+    output_file = fixture_path + function_name + '.npz'
+    
+    lp = -5
+    hp = 1.5
+    steps = 20
+    s_values = np.concatenate([-np.logspace(hp, lp, steps),
+                               [0],
+                               np.logspace(lp, hp, steps)])
+    sigmas = np.linspace(1, 100, 10)
+    
+    s_values, sigmas = np.meshgrid(s_values, sigmas)
+    s_values = s_values.flatten()
+    sigmas = sigmas.flatten()
+    
+    outputs = []
+    for s, sigma in zip(s_values, sigmas):
+        outputs.append(Phi_prime_mu(s, sigma))
+    
+    np.savez(output_file, s_values=s_values, sigmas=sigmas,
+             outputs=outputs)
+
+    
+def fix_power_spectra(network, file):
     network.power_spectra()
+    network.save(file_name=file)
+    
+
+def fix_Psi(fixture_path):
+    """Call Psi for a range of possible inputs and save result as fixture."""
+
+    function_name = 'Psi'
+    output_file = fixture_path + function_name + '.npz'
+    
+    z_range = np.concatenate([-np.logspace(2, -5, 4), [0],
+                              np.logspace(-5, 2, 4)])
+    a, b = np.meshgrid(z_range, z_range)
+    zs = a.flatten() + complex(0, 1) * b.flatten()
+    xs = np.linspace(-10, 10, 8)
+    
+    zs, xs = np.meshgrid(zs, xs)
+    zs = zs.flatten()
+    xs = xs.flatten()
+    
+    pcfu_outputs = []
+    for z, x in zip(zs, xs):
+        pcfu_outputs.append(mpmath.pcfu(z, -x))
+
+    outputs = []
+    for z, x in zip(zs, xs):
+        outputs.append(Psi(z, x))
+    
+    np.savez(output_file, zs=zs, xs=xs, pcfus=pcfu_outputs,
+             outputs=outputs)
     
     
-def fixture_sensitivity_measure(network):
+def fix_sensitivity_measure(network, file):
     omega = network.analysis_params['omega']
     network.sensitivity_measure(omega)
     network.transfer_function(omega)
+    network.save(file_name=file)
     
 
-def fixture_additional_rates_for_fixed_input(network):
-    nu_e_ext, nu_i_ext = network.additional_rates_for_fixed_input(
-        network.network_params['mean_input_set'],
-        network.network_params['std_input_set'])
-    network.results['add_nu_e_ext'] = nu_e_ext
-    network.results['add_nu_i_ext'] = nu_i_ext
+def fix_transfer_function(network, file):
+    """Calculate results for all options of transfer_function."""
+    network.transfer_function(method='shift')
+    network.results['tf_shift'] = network.results.pop('transfer_function')
+    network.transfer_function(method='taylor')
+    network.results['tf_taylor'] = network.results['transfer_function']
+    network.save(file_name=file)
     
 
-def fixture_eff_coupling_strength(network):
-    eff_coupling_strength = lmt.meanfield_calcs.effective_coupling_strength(
-        network.network_params['tau_m'],
-        network.network_params['tau_s'],
-        network.network_params['tau_r'],
-        network.network_params['V_0_rel'],
-        network.network_params['V_th_rel'],
-        network.network_params['J'],
-        network.results['mean_input'],
-        network.results['std_input'])
-    network.results['effective_coupling_strength'] = eff_coupling_strength
-
-
-def fixture_d_nu_d_mu(network):
-    params = network.network_params
-    mus = network.mean_input()
-    sigmas = network.std_input()
-    d_nu_d_mu = [lmt.aux_calcs.d_nu_d_mu(params['tau_m'],
-                                         params['tau_r'],
-                                         params['V_th_rel'],
-                                         params['V_0_rel'],
-                                         mu, sigma)
-                 for mu, sigma in zip(mus, sigmas)]
-    network.results['d_nu_d_mu'] = d_nu_d_mu
-
-
-def fixture_d_nu_d_mu_fb433(network):
-    params = network.network_params
-    mus = network.mean_input()
-    sigmas = network.std_input()
-    d_nu_d_mu_fb433 = [lmt.aux_calcs.d_nu_d_mu_fb433(params['tau_m'],
-                                                     params['tau_s'],
-                                                     params['tau_r'],
-                                                     params['V_th_rel'],
-                                                     params['V_0_rel'],
-                                                     mu, sigma)
-                       for mu, sigma in zip(mus, sigmas)]
-    network.results['d_nu_d_mu_fb433'] = d_nu_d_mu_fb433
-
-
-def fixture_d_nu_d_nu_in_fb(network):
-    params = network.network_params
-    mus = network.mean_input()
-    sigmas = network.std_input()
-    d_nu_d_nu_in_fb = [lmt.aux_calcs.d_nu_d_nu_in_fb(params['tau_m'],
-                                                     params['tau_s'],
-                                                     params['tau_r'],
-                                                     params['V_th_rel'],
-                                                     params['V_0_rel'],
-                                                     params['j'],
-                                                     mu, sigma)
-                       for mu, sigma in zip(mus, sigmas)]
-    network.results['d_nu_d_nu_in_fb'] = d_nu_d_nu_in_fb
+def fix_working_point(network, file):
+    network.working_point()
+    network.save(file_name=file)
 
 
 if __name__ == '__main__':
@@ -159,27 +347,43 @@ if __name__ == '__main__':
 
     # only run code if users are sure they want to do it
     if '--force' in args.keys():
-        fixture_path = 'tests/fixtures/data/'
-        configs = dict(noise_driven='network_params_microcircuit.yaml',
-                       negative_firing_rate='minimal_negative.yaml',
-                       # mean_driven='small_network.yaml',
-                       )
-        analysis_param_file = 'analysis_params_test.yaml'
 
+        fixture_path = 'tests/fixtures/data/'
+        config_path = 'tests/fixtures/config/'
+        
+        # purely numerical fixtures, that don't need any network
+        fix_Phi(fixture_path)
+        fix_Phi_prime_mu(fixture_path)
+        fix_Psi(fixture_path)
+        fix_d_Psi(fixture_path)
+        fix_d_2_Psi(fixture_path)
+        fix_p_hat_boxcar(fixture_path)
+        
+        # fixtures that need a network, or network params to be calculated
+        configs = dict(noise_driven=(config_path
+                                     + 'network_params_microcircuit.yaml'),
+                       negative_firing_rate=(config_path
+                                             + 'minimal_negative.yaml'),
+                       # mean_driven=(config_path
+                       #              + 'small_network.yaml'),
+                       )
+        analysis_param_file = config_path + 'analysis_params_test.yaml'
         for regime, param_file in configs.items():
+            
+            file_path = '{}{}_regime.h5'.format(fixture_path, regime)
+            
             network = lmt.Network(param_file, analysis_param_file)
             network.network_params['regime'] = regime
             
-            fixture_working_point(network)
-            fixture_transfer_function(network)
-            fixture_delay_dist_matrix(network)
-            fixture_sensitivity_measure(network)
-            fixture_power_spectra(network)
-            fixture_additional_rates_for_fixed_input(network)
-            fixture_eff_coupling_strength(network)
-            fixture_d_nu_d_mu(network)
-            fixture_d_nu_d_mu_fb433(network)
-            fixture_d_nu_d_nu_in_fb(network)
-        
-            network.save(file_name='{}{}_regime.h5'.format(fixture_path,
-                                                           regime))
+            fix_working_point(network, file_path)
+            fix_transfer_function(network, file_path)
+            fix_delay_dist_single(network, file_path)
+            fix_delay_dist_matrix(network, file_path)
+            fix_sensitivity_measure(network, file_path)
+            fix_power_spectra(network, file_path)
+            fix_eigenspectra(network, file_path)
+            fix_additional_rates_for_fixed_input(network, file_path)
+            fix_eff_coupling_strength(network, file_path)
+            fix_d_nu_d_mu(network, file_path)
+            fix_d_nu_d_mu_fb433(network, file_path)
+            fix_d_nu_d_nu_in_fb(network, file_path)
